Feat: 대표 이미지 추가 컴포넌트 제작 완료
