+++ src/components/Trip/TripPlan/TripPlanPostingPlan/TripPlanPostingPlan.tsx
@@ -1,14 +1,12 @@
 /* eslint-disable react/jsx-props-no-spreading */
 
-import React, { useRef, useState } from 'react';
+import React, { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { differenceInDays } from 'date-fns';
 import 'react-sliding-pane/dist/react-sliding-pane.css';
 import CalendarToday from '@mui/icons-material/CalendarMonth';
 import PlaceIcon from '@mui/icons-material/Place';
-import AddAPhotoIcon from '@mui/icons-material/AddAPhoto';
 import { SlArrowLeft } from 'react-icons/sl';
-import { IoCloseCircle } from 'react-icons/io5';
 import { GoPlusCircle } from 'react-icons/go';
 import TripPlanGoogleMaps from './TripPlanGoogleMaps/TripPlanGoogleMaps';
 import * as Styled from './TripPlanPostingPlan.styles';
@@ -17,13 +15,14 @@ import TripPlanPlaceModal from '../TripPlanPostingReview/TripPlanAddPlace/TripPl
 import TripPlanAddHashtags from '../TripPlanPostingReview/TripPlanAddHashtags/TripPlanAddHashtags';
 import TripPlanSetBudget from '../TripPlanPostingReview/TripPlanSetBudget/TripPlanSetBudget';
 import TripPlanAddTagsButton from '../TripPlanPostingReview/TripPlanAddTags/TripPlanAddTagsButton/TripPlanAddTagsButton';
+import TripPlanUploadImages from './TripPlanUploadImages/TripPlanUploadImages';
+import TripPlanUploadMainImages from './TripPlanUploadMainImages/TripPlanUploadMainImages';
 
 const TripPlanPosting = () => {
   const [selectedDay, setSelectedDay] = useState<number | null>(1);
   const [selectedPlace, setSelectedPlace] = useState<string>('');
   const { register, handleSubmit, setValue } = useForm();
-  const UploadImageIconRef = useRef<HTMLInputElement>(null);
-  const [uploadedImages, setUploadedImages] = useState<string[]>([]);
+
   const [isPlaceModalOpen, setIsPlaceModalOpen] = useState({
     isPaneOpenLeft: false,
   });
@@ -131,35 +130,6 @@ const TripPlanPosting = () => {
     //   });
   };
 
-  const handleUploadImage = (event: React.MouseEvent) => {
-    event.preventDefault();
-    UploadImageIconRef.current?.click();
-  };
-
-  const handleChangeImage = async (
-    event: React.ChangeEvent<HTMLInputElement>,
-    day: number,
-  ) => {
-    const file = event.target.files ? event.target.files[0] : null;
-    if (file) {
-      const fileUrl = URL.createObjectURL(file);
-      setUploadedImages((prevImages) => [...prevImages, fileUrl]);
-
-      setFormData((prevFormData) => {
-        const newFormData = [...prevFormData];
-        const firstPlace = newFormData[day].places[0];
-        newFormData[day].places[0] = { ...firstPlace, photo: fileUrl };
-        return newFormData;
-      });
-    }
-  };
-
-  const handleRemoveImage = (imageNumber: number) => {
-    setUploadedImages((currentImages) =>
-      currentImages.filter((_, index) => index !== imageNumber),
-    );
-  };
-
   const createDaysInput = () => {
     return (
       <>
@@ -233,45 +203,10 @@ const TripPlanPosting = () => {
               }
             }}
           />
-
-          <Styled.UploadImageContainer>
-            <Styled.UploadImageIcon onClick={handleUploadImage}>
-              <AddAPhotoIcon className="image-icon" />
-              <p className="photo-text">사진 업로드</p>
-            </Styled.UploadImageIcon>
-
-            <Styled.ImageInput
-              ref={UploadImageIconRef}
-              type="file"
-              accept="image/*"
-              name="file"
-              onChange={(event) =>
-                selectedDay !== null &&
-                handleChangeImage(event, selectedDay - 1)
-              }
-              disabled={uploadedImages.length >= 5}
-            />
-
-            <Styled.UploadedImageSwiper
-              spaceBetween={1}
-              slidesPerView={uploadedImages.length > 1 ? 1.5 : 1}
-              direction="horizontal"
-              scrollbar={{
-                draggable: true,
-                el: '.swiper-scrollbar',
-                hide: false,
-              }}>
-              {uploadedImages.map((image, number) => (
-                <Styled.UploadedImage key={image}>
-                  <img src={image} alt={`Uploaded ${number + 1}`} />
-                  <Styled.RemoveBtn onClick={() => handleRemoveImage(number)}>
-                    <IoCloseCircle color="#626161" />
-                  </Styled.RemoveBtn>
-                </Styled.UploadedImage>
-              ))}
-            </Styled.UploadedImageSwiper>
-          </Styled.UploadImageContainer>
-
+          <TripPlanUploadImages
+            setFormData={setFormData}
+            selectedDay={selectedDay}
+          />
           <TripPlanAddTagsButton />
         </Styled.PostingForm>
         <TripPlanSetBudget />
@@ -290,6 +225,8 @@ const TripPlanPosting = () => {
           <div className="nightndays">{getNightAndDays()}</div>
           <CalendarToday className="calendar-icon" />
         </Styled.DateDisplay>
+
+        <TripPlanUploadMainImages setFormData={setFormData} />
 
         <Styled.GoogleMapsContainer>
           <TripPlanGoogleMaps />
