+++ src/components/Trip/TripPlan/TripPlanPostingPlan/TripPlanUploadImages/TripPlanUploadImages.types.ts
@@ -1,0 +1,4 @@
+export interface UploadImagesProps {
+  setFormData: (formData: any) => void; // any 추후 타입 수정 필요
+  selectedDay?: number | null;
+}
